Yahooooo~!
Cleanup
Finished Medkit&Battery Item&Entity~

+++ src/cbproject/core/renderers/RenderUtils.java
@@ -2,7 +2,6 @@ package cbproject.core.renderers;
 
 import org.lwjgl.opengl.GL11;
 
-import net.minecraft.block.Block;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.util.Vec3;
 import cpw.mods.fml.client.FMLClientHandler;
@@ -135,34 +134,34 @@ public class RenderUtils{
     {
         par0Tessellator.startDrawingQuads();
         par0Tessellator.setNormal(0.0F, 0.0F, 1.0F);
-        par0Tessellator.addVertexWithUV(0.0D, 0.0D, 0.0D, (double)par1, (double)par4);
-        par0Tessellator.addVertexWithUV(1.0D, 0.0D, 0.0D, (double)par3, (double)par4);
-        par0Tessellator.addVertexWithUV(1.0D, 1.0D, 0.0D, (double)par3, (double)par2);
-        par0Tessellator.addVertexWithUV(0.0D, 1.0D, 0.0D, (double)par1, (double)par2);
+        par0Tessellator.addVertexWithUV(0.0D, 0.0D, 0.0D, par1, par4);
+        par0Tessellator.addVertexWithUV(1.0D, 0.0D, 0.0D, par3, par4);
+        par0Tessellator.addVertexWithUV(1.0D, 1.0D, 0.0D, par3, par2);
+        par0Tessellator.addVertexWithUV(0.0D, 1.0D, 0.0D, par1, par2);
         par0Tessellator.draw();
         par0Tessellator.startDrawingQuads();
         par0Tessellator.setNormal(0.0F, 0.0F, -1.0F);
-        par0Tessellator.addVertexWithUV(0.0D, 1.0D, (double)(0.0F - par7), (double)par1, (double)par2);
-        par0Tessellator.addVertexWithUV(1.0D, 1.0D, (double)(0.0F - par7), (double)par3, (double)par2);
-        par0Tessellator.addVertexWithUV(1.0D, 0.0D, (double)(0.0F - par7), (double)par3, (double)par4);
-        par0Tessellator.addVertexWithUV(0.0D, 0.0D, (double)(0.0F - par7), (double)par1, (double)par4);
+        par0Tessellator.addVertexWithUV(0.0D, 1.0D, 0.0F - par7, par1, par2);
+        par0Tessellator.addVertexWithUV(1.0D, 1.0D, 0.0F - par7, par3, par2);
+        par0Tessellator.addVertexWithUV(1.0D, 0.0D, 0.0F - par7, par3, par4);
+        par0Tessellator.addVertexWithUV(0.0D, 0.0D, 0.0F - par7, par1, par4);
         par0Tessellator.draw();
-        float f5 = (float)par5 * (par1 - par3);
-        float f6 = (float)par6 * (par4 - par2);
+        float f5 = par5 * (par1 - par3);
+        float f6 = par6 * (par4 - par2);
         par0Tessellator.startDrawingQuads();
         par0Tessellator.setNormal(-1.0F, 0.0F, 0.0F);
         int k;
         float f7;
         float f8;
 
-        for (k = 0; (float)k < f5; ++k)
+        for (k = 0; k < f5; ++k)
         {
-            f7 = (float)k / f5;
-            f8 = par1 + (par3 - par1) * f7 - 0.5F / (float)par5;
-            par0Tessellator.addVertexWithUV((double)f7, 0.0D, (double)(0.0F - par7), (double)f8, (double)par4);
-            par0Tessellator.addVertexWithUV((double)f7, 0.0D, 0.0D, (double)f8, (double)par4);
-            par0Tessellator.addVertexWithUV((double)f7, 1.0D, 0.0D, (double)f8, (double)par2);
-            par0Tessellator.addVertexWithUV((double)f7, 1.0D, (double)(0.0F - par7), (double)f8, (double)par2);
+            f7 = k / f5;
+            f8 = par1 + (par3 - par1) * f7 - 0.5F / par5;
+            par0Tessellator.addVertexWithUV(f7, 0.0D, 0.0F - par7, f8, par4);
+            par0Tessellator.addVertexWithUV(f7, 0.0D, 0.0D, f8, par4);
+            par0Tessellator.addVertexWithUV(f7, 1.0D, 0.0D, f8, par2);
+            par0Tessellator.addVertexWithUV(f7, 1.0D, 0.0F - par7, f8, par2);
         }
 
         par0Tessellator.draw();
@@ -170,44 +169,44 @@ public class RenderUtils{
         par0Tessellator.setNormal(1.0F, 0.0F, 0.0F);
         float f9;
 
-        for (k = 0; (float)k < f5; ++k)
+        for (k = 0; k < f5; ++k)
         {
-            f7 = (float)k / f5;
-            f8 = par1 + (par3 - par1) * f7 - 0.5F / (float)par5;
+            f7 = k / f5;
+            f8 = par1 + (par3 - par1) * f7 - 0.5F / par5;
             f9 = f7 + 1.0F / f5;
-            par0Tessellator.addVertexWithUV((double)f9, 1.0D, (double)(0.0F - par7), (double)f8, (double)par2);
-            par0Tessellator.addVertexWithUV((double)f9, 1.0D, 0.0D, (double)f8, (double)par2);
-            par0Tessellator.addVertexWithUV((double)f9, 0.0D, 0.0D, (double)f8, (double)par4);
-            par0Tessellator.addVertexWithUV((double)f9, 0.0D, (double)(0.0F - par7), (double)f8, (double)par4);
+            par0Tessellator.addVertexWithUV(f9, 1.0D, 0.0F - par7, f8, par2);
+            par0Tessellator.addVertexWithUV(f9, 1.0D, 0.0D, f8, par2);
+            par0Tessellator.addVertexWithUV(f9, 0.0D, 0.0D, f8, par4);
+            par0Tessellator.addVertexWithUV(f9, 0.0D, 0.0F - par7, f8, par4);
         }
 
         par0Tessellator.draw();
         par0Tessellator.startDrawingQuads();
         par0Tessellator.setNormal(0.0F, 1.0F, 0.0F);
 
-        for (k = 0; (float)k < f6; ++k)
+        for (k = 0; k < f6; ++k)
         {
-            f7 = (float)k / f6;
-            f8 = par4 + (par2 - par4) * f7 - 0.5F / (float)par6;
+            f7 = k / f6;
+            f8 = par4 + (par2 - par4) * f7 - 0.5F / par6;
             f9 = f7 + 1.0F / f6;
-            par0Tessellator.addVertexWithUV(0.0D, (double)f9, 0.0D, (double)par1, (double)f8);
-            par0Tessellator.addVertexWithUV(1.0D, (double)f9, 0.0D, (double)par3, (double)f8);
-            par0Tessellator.addVertexWithUV(1.0D, (double)f9, (double)(0.0F - par7), (double)par3, (double)f8);
-            par0Tessellator.addVertexWithUV(0.0D, (double)f9, (double)(0.0F - par7), (double)par1, (double)f8);
+            par0Tessellator.addVertexWithUV(0.0D, f9, 0.0D, par1, f8);
+            par0Tessellator.addVertexWithUV(1.0D, f9, 0.0D, par3, f8);
+            par0Tessellator.addVertexWithUV(1.0D, f9, 0.0F - par7, par3, f8);
+            par0Tessellator.addVertexWithUV(0.0D, f9, 0.0F - par7, par1, f8);
         }
 
         par0Tessellator.draw();
         par0Tessellator.startDrawingQuads();
         par0Tessellator.setNormal(0.0F, -1.0F, 0.0F);
 
-        for (k = 0; (float)k < f6; ++k)
+        for (k = 0; k < f6; ++k)
         {
-            f7 = (float)k / f6;
-            f8 = par4 + (par2 - par4) * f7 - 0.5F / (float)par6;
-            par0Tessellator.addVertexWithUV(1.0D, (double)f7, 0.0D, (double)par3, (double)f8);
-            par0Tessellator.addVertexWithUV(0.0D, (double)f7, 0.0D, (double)par1, (double)f8);
-            par0Tessellator.addVertexWithUV(0.0D, (double)f7, (double)(0.0F - par7), (double)par1, (double)f8);
-            par0Tessellator.addVertexWithUV(1.0D, (double)f7, (double)(0.0F - par7), (double)par3, (double)f8);
+            f7 = k / f6;
+            f8 = par4 + (par2 - par4) * f7 - 0.5F / par6;
+            par0Tessellator.addVertexWithUV(1.0D, f7, 0.0D, par3, f8);
+            par0Tessellator.addVertexWithUV(0.0D, f7, 0.0D, par1, f8);
+            par0Tessellator.addVertexWithUV(0.0D, f7, 0.0F - par7, par1, f8);
+            par0Tessellator.addVertexWithUV(1.0D, f7, 0.0F - par7, par3, f8);
         }
 
         par0Tessellator.draw();
